osb.c: Implemented OSB

+++ src/vc/bow.c
@@ -14,7 +14,7 @@ int main(){
     unsigned char *phrase;
     size_t size_phrase = 0;
     ssize_t len_phrase;
-    unsigned long hash_result;
+    uint32_t hash_result;
 
     while ((len_phrase = getline((char **)(&phrase), &size_phrase, stdin)) != -1){
         strip_endl(phrase, &len_phrase);
