+++ src/vc/ngram.c
@@ -33,7 +33,7 @@ int main(){
     size_t ptr;  /* Pointer used when concatenating phrases. */
     unsigned int not_null_phrase = 0;
 
-    unsigned long hash;
+    uint32_t hash;
 
     while ((len_phrase = getline((char **)(&phrase), &size_phrase, stdin)) != -1){
         strip_endl(phrase, &len_phrase);
@@ -82,13 +82,13 @@ int main(){
 
         if (not_null_phrase < PHRASE_PER_GROUP){
             len_catphrase = 0;
-            for (unsigned i = not_null_phrase; i < PHRASE_PER_GROUP; i++){
+            for (size_t i = not_null_phrase; i < PHRASE_PER_GROUP; i++){
                 len_catphrase += len_phrase_list[i] + 1;
             }
             catphrase = malloc(len_catphrase-- * sizeof(char));
 
             ptr = 0;
-            for (unsigned int i = not_null_phrase; i < PHRASE_PER_GROUP; i++){
+            for (size_t i = not_null_phrase; i < PHRASE_PER_GROUP; i++){
                 memcpy(catphrase + ptr * sizeof(char), phrase_list[i], len_phrase_list[i]);
                 ptr += len_phrase_list[i];
                 catphrase[ptr++] = '\0';
@@ -101,7 +101,7 @@ int main(){
         }
     }
 
-    for (unsigned int i = 0; i < PHRASE_PER_GROUP; i++){
+    for (size_t i = 0; i < PHRASE_PER_GROUP; i++){
         free(phrase_list[i]);
     }
     free(phrase);
